RFC-0038: the std functions that refuse an input return a Result
A failure the program can act on is a value: base64_decode returns
Result<String, Base64Error>, char_to_int and int_to_char return
Result<_, CharError>, and parsing text is not a conversion: each integer
type has its own from_str under its namespace, u64::from_str(s), returning
Result<T, ParseIntError>. core::to_int keeps only the instances whose every
value is an i64, so the String and u64 instances are gone and its return
is a plain i64. A count that must not be negative is a u64: repeat_str,
take, and skip take one, and the Iter pipeline counts in u64.

decimal_to_float returns f64: rust_decimal 1.40.0's to_f64 is Some on
every path, so the Option was the trait's shape and not a failure.

regex, parse_date, from_timestamp, and decimal stay at Result<_, Trap>:
Result<ExternType, E> crosses as an opaque Rust value because
acvus-extern's Cross impl for Result requires T: Cross and an extension
type has none; the fix belongs in acvus-extern.

+++ acvus-ext/src/iter_pipeline.rs
@@ -37,8 +37,8 @@ type Lambda<Rt> = Fn1<<Rt as Runtime>::Value, <Rt as Runtime>::Value, (), Rt>;
 enum Op<Rt: Runtime> {
     Map(Lambda<Rt>),
     Filter(Lambda<Rt>),
-    Take { remaining: usize },
-    Skip { remaining: usize },
+    Take { remaining: u64 },
+    Skip { remaining: u64 },
     Flatten(Expand<Rt>),
     FlatMap(Lambda<Rt>, Expand<Rt>),
 }
@@ -155,11 +155,11 @@ where
         Self::erased(self.0.push_op(Op::Filter(f.erased())))
     }
 
-    pub fn take(self, n: usize) -> Self {
+    pub fn take(self, n: u64) -> Self {
         Self::erased(self.0.push_op(Op::Take { remaining: n }))
     }
 
-    pub fn skip(self, n: usize) -> Self {
+    pub fn skip(self, n: u64) -> Self {
         Self::erased(self.0.push_op(Op::Skip { remaining: n }))
     }
